# docs/web-control-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChatWire local moderator API",
    "version": "1.0.0",
    "description": "Private moderator dashboard API. Issue browser access via the Discord /web command. No public endpoint issues grants. See web-control-setup.md for deployment and pending settings semantics."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "session": []
    }
  ],
  "paths": {
    "/api/v1/me": {
      "get": {
        "summary": "Current moderator identity and CSRF token",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/servers": {
      "get": {
        "summary": "All registered instance summaries",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/servers/{server}": {
      "get": {
        "summary": "Runtime state, boot ID, and desired/runtime config revisions",
        "parameters": [
          {
            "name": "server",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/settings/global": {
      "get": {
        "summary": "Read redacted desired settings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Persist a validated settings patch; explicit reload/restart applies it",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Obtain from GET /api/v1/me; Origin must equal the configured public origin."
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Quoted settings revision."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/global/schema": {
      "get": {
        "summary": "Settings metadata and validation constraints",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Field"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/host/actions/{action}": {
      "post": {
        "summary": "Submit confirmed action",
        "parameters": [
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "factorio-start",
                "factorio-stop",
                "factorio-restart",
                "chatwire-restart",
                "config-reload",
                "map-create",
                "map-reset",
                "map-load",
                "map-exchange",
                "map-archive",
                "mods-update",
                "mods-sync",
                "mods-edit",
                "mods-clear",
                "mods-clear-history",
                "factorio-update",
                "factorio-install",
                "rcon",
                "upload-apply",
                "player-level",
                "ip-ban",
                "ip-unban"
              ]
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Obtain from GET /api/v1/me; Origin must equal the configured public origin."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 80
            },
            "description": "Reuse the same key and body after a lost response; do not generate a replacement key."
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionParameters"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/servers/{server}/players": {
      "get": {
        "summary": "Online players in this instance",
        "parameters": [
          {
            "name": "server",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/servers/{server}/jobs": {
      "get": {
        "summary": "Latest 100 instance jobs",
        "parameters": [
          {
            "name": "server",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/servers/{server}/audit": {
      "get": {
        "summary": "Recent instance action records",
        "parameters": [
          {
            "name": "server",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/servers/{server}/uploads": {
      "post": {
        "summary": "Stage one upload without activating it",
        "parameters": [
          {
            "name": "server",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Obtain from GET /api/v1/me; Origin must equal the configured public origin."
          }
        ],
        "responses": {
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Upload"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "minProperties": 1,
                "maxProperties": 1,
                "additionalProperties": false,
                "properties": {
                  "save": {
                    "type": "string",
                    "format": "binary"
                  },
                  "mod-list": {
                    "type": "string",
                    "format": "binary"
                  },
                  "mod-settings": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/players": {
      "get": {
        "summary": "Shared player database, paginated",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/players/{player}": {
      "get": {
        "summary": "Shared player record without linked Discord ID",
        "parameters": [
          {
            "name": "player",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/host/ip-bans": {
      "get": {
        "summary": "Public IPv4 firewall deny entries",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs": {
      "get": {
        "summary": "Latest jobs grouped by instance plus host batches",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job}": {
      "get": {
        "summary": "Read a host batch job",
        "parameters": [
          {
            "name": "job",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit": {
      "get": {
        "summary": "Recent audit records grouped by instance and host",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events": {
      "get": {
        "summary": "SSE snapshot invalidation; refetch snapshots on resync_required",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-chatwire"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Settings": {
        "type": "object",
        "properties": {
          "revision": {
            "type": "string"
          },
          "values": {
            "type": "object",
            "additionalProperties": true
          },
          "application": {
            "type": "string"
          }
        }
      },
      "Field": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "admin": {
            "type": "boolean"
          },
          "secret": {
            "type": "boolean"
          },
          "read_only": {
            "type": "boolean"
          },
          "effect": {
            "type": "string"
          },
          "min": {
            "type": "number"
          },
          "max": {
            "type": "number"
          },
          "choices": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "label": {
            "type": "string"
          }
        }
      },
...
      "Upload": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "save",
              "mod-list",
              "mod-settings"
            ]
          },
          "name": {
            "type": "string"
          },
          "size": {
            "type": "integer"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}